+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Breaking
+
+* *Breaking*: Removed `Gewerk\RecurringDates\Model\Occurrence::fromArray()`, use `new Occurrence($config)` instead
+
 ## 1.0.0-alpha.2 - 2023-12-05
 
 ### Breaking